pyproject.toml
[tool.poetry]
name = "solaris pygame"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.16.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
